feat: remove extra spacing in pokedex page
