+++ pages/pokedex/index.tsx
@@ -1,0 +1,59 @@
+import Link from "next/link";
+
+const Pokédex = () => {
+  return (
+    <>
+      <section className="text-gray-700 body-font">
+        <div className="container px-5 py-10 mx-auto">
+          <div className="flex flex-wrap -m-4">
+            <Link href="/pokedex/a">
+              <div className="xl:w-1/4 md:w-1/2 p-4">
+                <div className="bg-gray-100 p-6 rounded-lg">
+                  <img
+                    className="h-40 rounded w-full object-cover object-center mb-6"
+                    src="https://dummyimage.com/720x400"
+                    alt="content"
+                  />
+                  <h3 className="tracking-widest text-red-500 text-xs font-medium title-font">
+                    SUBTITLE
+                  </h3>
+                  <h2 className="text-lg text-gray-900 font-medium title-font mb-4">
+                    Chichen Itza
+                  </h2>
+                  <p className="leading-relaxed text-base">
+                    Fingerstache flexitarian street art 8-bit waistcoat.
+                    Distillery hexagon disrupt edison bulbche.
+                  </p>
+                </div>
+              </div>
+            </Link>
+
+            <Link href="/pokedex/a">
+              <div className="xl:w-1/4 md:w-1/2 p-4">
+                <div className="bg-gray-100 p-6 rounded-lg">
+                  <img
+                    className="h-40 rounded w-full object-cover object-center mb-6"
+                    src="https://dummyimage.com/720x400"
+                    alt="content"
+                  />
+                  <h3 className="tracking-widest text-red-500 text-xs font-medium title-font">
+                    SUBTITLE
+                  </h3>
+                  <h2 className="text-lg text-gray-900 font-medium title-font mb-4">
+                    Chichen Itza
+                  </h2>
+                  <p className="leading-relaxed text-base">
+                    Fingerstache flexitarian street art 8-bit waistcoat.
+                    Distillery hexagon disrupt edison bulbche.
+                  </p>
+                </div>
+              </div>
+            </Link>
+          </div>
+        </div>
+      </section>
+    </>
+  );
+};
+
+export default Pokédex;
